requirements.txt
Authlib==0.14.3
django==4.2
django-debug-toolbar==4.0.0
django-widget-tweaks==1.4.12
flake8==3.8.3
flake8-isort==3.0.0
